code review version 8
